requirement.txt
streamlit
pandas
numpy
scikit-learn
Pillow
requests